TST: Move tests around to match new structure
