+++ openff/interchange/tests/unit_tests/smirnoff/test_create.py
@@ -1,0 +1,5 @@
+from openff.interchange.tests import _BaseTest
+
+
+class TestCreate(_BaseTest):
+    pass
